Remove support for reading from files
On second thought, we don't really need this because it already reads
from stdin, so can use head/tail/etc. I originally added this to support
the Vim integration, but that just passes stdin too now.

Problem with current integration is that it doesn't display errors in
many cases (e.g. file doesn't exist), and adding them would be hard in
cases where there is an actual :# in the string.